manuscript-extractor/requirements.txt
PyMuPDF>=1.23.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
unstructured[pdf]>=0.10.0
